requirements.txt
evernote3
lxml~=4.9.2
oauth2
openai~=0.27.2
requests~=2.28.2
beautifulsoup4~=4.12.2